red_blue/client/pyyaml-vuln/requirements.txt
Flask
PyYAML==5.1
